# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Petlove API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/concorrentes/": {
      "get": {
        "tags": [
          "Concorrentes"
        ],
        "description": "Endpoint para buscar todos os concorrentes.",
        "responses": {
          "200": {
            "description": "Lista de concorrentes encontrada com sucesso.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "nome": {
                    "type": "string",
                    "example": "Petz"
                  },
                  "tipo": {
                    "type": "string",
                    "example": "Petshop"
                  },
                  "endereco": {
                    "type": "string",
                    "example": "Rua Exemplo, 123"
                  },
                  "foto": {
                    "type": "string",
                    "example": "http://localhost:8000/uploads/concorrentes/foto.jpg"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00.000Z"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00.000Z"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Concorrentes"
        ],
        "description": "Endpoint para criar um novo concorrente.",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "description": "Foto do concorrente."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "foto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Concorrente criado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Concorrente criado com sucesso"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/concorrentes/{id}": {
      "get": {
        "tags": [
          "Concorrentes"
        ],
        "description": "Endpoint para buscar um concorrente pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Concorrente encontrado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "Petz"
                },
                "tipo": {
                  "type": "string",
                  "example": "Petshop"
                },
                "endereco": {
                  "type": "string",
                  "example": "Rua Exemplo, 123"
                },
                "foto": {
                  "type": "string",
                  "example": "http://localhost:8000/uploads/concorrentes/foto.jpg"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023-10-27T10:00:00.000Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2023-10-27T10:00:00.000Z"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Concorrentes"
        ],
        "description": "Endpoint para editar um concorrente existente.",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "description": "Foto do concorrente."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "foto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Concorrente editado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Concorrente atualizado com sucesso"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Concorrentes"
        ],
        "description": "Endpoint para deletar um concorrente pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Concorrente deletado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Concorrente deletado com sucesso"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/produtos/": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para buscar todos os produtos.",
        "responses": {
          "200": {
            "description": "Lista de produtos encontrada com sucesso.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "nome": {
                    "type": "string",
                    "example": "Nome do Produto"
                  },
                  "preco": {
                    "type": "number",
                    "example": 10.5
                  },
                  "codigo": {
                    "type": "string",
                    "example": "COD123"
                  },
                  "foto": {
                    "type": "string",
                    "example": "http://localhost:8000/uploads/produtos/foto.jpg"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00.000Z"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00.000Z"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para criar um novo produto.",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "nome",
            "in": "formData",
            "type": "string",
            "description": "Nome do produto."
          },
          {
            "name": "preco",
            "in": "formData",
            "type": "number",
            "description": "Preço do produto."
          },
          {
            "name": "codigo",
            "in": "formData",
            "type": "string",
            "description": "Código do produto."
          },
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "description": "Foto do produto."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "preco": {
                  "example": "any"
                },
                "foto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto criado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Produto criado com sucesso"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/produtos/{id}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para buscar um produto pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "Nome do Produto"
                },
                "preco": {
                  "type": "number",
                  "example": 10.5
                },
                "codigo": {
                  "type": "string",
                  "example": "COD123"
                },
                "foto": {
                  "type": "string",
                  "example": "http://localhost:8000/uploads/produtos/foto.jpg"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023-10-27T10:00:00.000Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2023-10-27T10:00:00.000Z"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para editar um produto existente.",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "nome",
            "in": "formData",
            "type": "string",
            "description": "Nome do produto."
          },
          {
            "name": "preco",
            "in": "formData",
            "type": "number",
            "description": "Preço do produto."
          },
          {
            "name": "codigo",
            "in": "formData",
            "type": "string",
            "description": "Código do produto."
          },
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "description": "Foto do produto."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "preco": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                },
                "foto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto editado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Produto atualizado com sucesso"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Produtos"
        ],
        "description": "Endpoint para deletar um produto pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto deletado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Produto deletado com sucesso"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}